refactor(fe): change mystudies to mystudy
